compositor|refactor: Split the snapshot-fit assertion into its two refusals
Extracts `assertSourcesFit` into `snapshot/`, carrying the refusal of a config whose adopted sources have moved since the snapshot was captured. `assertSnapshotFits` keeps its name, its `PlannableSnapshot` narrowing, and both callers, and now composes that check with the target-state requirement a plan needs.

Validate reports authoring faults over a snapshot captured without target state, so it needs the source-move refusal without the one that would turn such a snapshot away. Both still raise `StaleSnapshotError`, and both messages are unchanged.
